% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllIdentical.R
\name{AllIdentical}
\alias{AllIdentical}
\title{@title AllIdentical Function}
\usage{
AllIdentical(l)
}
\description{
@description Returns TRUE iff all arguments are identical.
}
\details{
@param A list of objects for which 'identical' is defined.

@author Robert Corty

@usage l <- list(a, b, c)
        AllIdentical(l))

@export
}

