% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArticlePrecedingNumber.R
\name{ArticlePrecedingNumber}
\alias{ArticlePrecedingNumber}
\title{ArticlePrecedingNumber Function}
\usage{
ArticlePrecedingNumber(n, definite = FALSE)
}
\arguments{
\item{n}{the number}

\item{definite}{whether the definite (or indefinite) article is
required.  Defaults to \code{FALSE}.}
}
\value{
a character object from \code{c('a', 'an', 'the')}
appropriate to precede the number
}
\description{
This function returns the appropriate english definite
or indefinite article to precede a number.  The possible return
values are 'a', 'an', and 'the'
}
\examples{
ArticlePrecedingNumber(-8)
ArticlePrecedingNumber(-2)
ArticlePrecedingNumber(2)
ArticlePrecedingNumber(8)
ArticlePrecedingNumber(8, definite = TRUE)

}
\author{
Robert Corty
}

