% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlankPlot.R
\name{BlankPlot}
\alias{BlankPlot}
\title{BlankPlot Function}
\usage{
BlankPlot(xlim = c(0, 1), ylim = c(0, 1), ...)
}
\arguments{
\item{xlim}{x axis limits for the plot, as in \code{plot}}

\item{ylim}{y axis limits for the plot, as in \code{plot}}

\item{...}{additional graphical parameters}
}
\value{
No return value.  Just draws the plot.
}
\description{
This function draws a plot using base \R
graphics with no axes, no axis labels, no title, and no data.
It should usually be used in conjunction with other functions for
plotting in base \R, such as \code{points}, \code{lines},
\code{rect}, etc.
}
\examples{
BlankPlot()
points(x = runif(n = 10), y = runif(n = 10))
BlankPlot(xlim = c(0, 100), ylim = c(0, 100))
lines(x = rnorm(n = 10, mean = 50, sd = 10),
      y = rnorm(n = 10, mean = 50, sd = 10))
BlankPlot(main = "My Great Plot")

}
\author{
Robert Corty
}
