% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsub_master.R
\name{bsub_master}
\alias{bsub_master}
\title{bsub_master Function}
\usage{
bsub_master(f, param.list, combination.type = "hypercube", q = "day",
  M = "5")
}
\arguments{
\item{f}{the function to be evaluated with potentially many different
parameter sets}

\item{param.list}{A named list of vectors of parameters.
By default, all combinations will be used.}
}
\value{
t
}
\description{

}
\details{

}
\examples{


}
\author{
Robert Corty
}

