% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonest_values.R
\name{commonest_values}
\alias{commonest_values}
\title{commonest_values}
\usage{
commonest_values(v, n = 5)
}
\arguments{
\item{v}{the vector to assess}

\item{n}{the number of common values are returned}
}
\value{
the commonest n values in v, sorted
}
