% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF2String.R
\name{DF2String}
\alias{DF2String}
\title{DF2String}
\usage{
DF2String(DF, signif)
}
\arguments{
\item{DF}{data.frame to be turned into string}

\item{signif}{number of significant digits to keep of numeric columns}
}
\value{
a string
}
\description{
DF2String
}
\examples{
DF2String(iris)
}
