% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEV.R
\name{DGEV}
\alias{DGEV}
\title{DGEV}
\usage{
DGEV(xs, gev, log = FALSE)
}
\arguments{
\item{xs}{the xs}

\item{gev}{the fitted gev}

\item{log}{whether to return log likelihood}
}
\value{
the density at xs
}
\description{
DGEV
}
