% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divvy.R
\name{Divvy}
\alias{Divvy}
\title{Divvy Function}
\usage{
Divvy(first.idx, last.idx, which.group,
  num.groups = ceiling(n.idxs/group.size),
  group.size = ceiling(n.idxs/num.groups))
}
\arguments{
\item{num.groups}{number of groups to divvy the items into, defaults to
\code{ceiling(n/group.size)}}

\item{group.size}{size of each group, defaults to
\code{ceiling(n/num.groups)}}

\item{n}{number of items to divvy up}

\item{i}{group of items to return}
}
\value{
the indices of the \code{i}th group of size \code{g} in \code{1:n}
}
\description{
This function splits the integers from 1 to \code{n} into
\code{g} groups and returns the \code{i}th group.  All of this is
straightforward, but it's important to do it in a consistent way
for splitting up data/analyses.
}
\details{
The \code{g}th group may be smaller than the other groups
}
\examples{


}
\author{
Robert Corty
}

