% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forfor.R
\name{forfor}
\alias{forfor}
\title{forfor}
\usage{
forfor(l, e)
}
\arguments{
\item{l}{named list of variables to loop over and their values to loop over}

\item{e}{expression to evaluate for every combination of values in l}
}
\details{
so  forfor(l = list(a = c(1,2), b = c(10, 20)), e = a + b)
}

