% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/into.R
\name{into}
\alias{into}
\title{into}
\usage{
into(value, name)
}
\arguments{
\item{value}{the value to store}

\item{name}{the expression to store it as}
}
\value{
nothing, just assigns
}
\details{
from Konrad Rudolph in a comment on the win-vector blog
}
