% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsBetween.R
\name{IsBetween}
\alias{IsBetween}
\title{IsBetween Function}
\usage{
IsBetween(x, lb, ub, equality.ok = FALSE)
}
\arguments{
\item{x}{the focal object}

\item{lb}{lower bound}

\item{ub}{upper bound}

\item{equality.ok}{Indicates whether it acceptable for the focal
object to be equal to one of the arguments.  Defaults to
\code{FALSE}.}
}
\value{
\code{TRUE} if \code{x} is between \code{lb} and
\code{ub} and \code{FALSE} otherwise.
}
\description{
This function returns \code{TRUE} if \code{x}
is between \code{lb} (lower bound) and \code{ub}
(upper bound).  This function can act on any type for which
\code{<, <=, >}, and \code{>=} are defined.  This function
works nicely with the pipe operator from the \pkg{dplyr} and
\pkg{magrittr} packages.
}
\examples{
IsBetween(x = 0, lb = 1, ub = 2)
IsBetween(x = 1, lb = 0, ub = 2)
IsBetween(x = 0, lb = 0, ub = 2)
IsBetween(x = 0, lb = 0, ub = 2, equality.ok = TRUE)


}
\author{
Robert Corty
}
