% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Take a Peek at Your Data

This is a variant on [tibble::glimpse()], that gives better
summaries of factor and character-type columns.}
\usage{
peek(x, width = NULL, ...)
}
\arguments{
\item{x}{An object to peek at.}

\item{width}{Width of output: defaults to the setting of the option
`tibble.width` (if finite) or the width of the console.}

\item{...}{Other arguments passed on to individual methods.}
}
\value{
x original x is (invisibly) returned, allowing `peek()` to be
  used within a data pipe line.
}
\section{S3 methods}{

`peek` is an S3 generic with a customised method for `tbl`s and
`data.frame`s, and a default method that calls [str()].
}

\examples{
peek(mtcars)

if (!requireNamespace("nycflights13", quietly = TRUE))
  stop("Please install the nycflights13 package to run the rest of this example")

peek(nycflights13::flights)
}
