% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_cyclic.R
\name{permute_cyclic}
\alias{permute_cyclic}
\title{permute a vector cyclically}
\usage{
permute_cyclic(v, n = 1)
}
\arguments{
\item{v}{the vector to permute}

\item{n}{the number of steps to cycle the vector}
}
\value{
the permuted vector
}
\description{
permute a vector cyclically
}
\examples{
cycle_n(v = 1:5)
cycle_n(v = 1:5, n = 3)
cycle_n(v = 1:5, n = 8)
}
