% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRWC.R
\name{saveRWC}
\alias{saveRWC}
\title{saveRWC}
\usage{
saveRWC(object, file_name, overwrite = FALSE, postfix_datetime = FALSE, ...)
}
\arguments{
\item{overwrite}{should the function overwrite an existing file?}

\item{postfix_datetime}{should the function postfix the save file name with the ISO-formatted datetime?}

\item{...}{}
}
\value{
TRUE is successful and FALSE otherwise
}
