% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servr_patch.R
\name{servr_patch}
\alias{servr_patch}
\title{servr_patch}
\usage{
servr_patch(path = "_site/")
}
\arguments{
\item{path}{the path in which to do the patching. Defaults to '_site'}
}
\value{
TRUE, if it completes successfully
}
\description{
\code{servr_patch} goes through all the html files in \code{path}
and replaces all instances of 'src="figure' with 'src="/figure'.  This is
a very kludgey way to fix a small bug in \pkg{servr}.
}

